list replies

+++ server/src/controllers/api/comment/delete-comment.ts
@@ -3,15 +3,18 @@ import * as _ from "lodash";
 import { IRemoveCommentById } from "../../../use-cases/comment/delete-comment-by-id";
 import { IGetCommentById } from "../../../use-cases/comment/get-comment-by-id";
 import { IDeleteCommentAssetByCommentId } from "../../../use-cases/comment-asset/delete-comment-asset-by-comment-id";
+import { IGetCommentAssetByCommentId } from "../../../use-cases/comment-asset/get-comment-asset-by-comment-id";
 
 export default function makeDeleteCommentController({
   deleteCommentById,
   getCommentById,
   deleteCommentAssetByCommentId,
+  getCommentAssetByCommentId,
 }: {
   deleteCommentById: IRemoveCommentById;
   getCommentById: IGetCommentById;
   deleteCommentAssetByCommentId: IDeleteCommentAssetByCommentId;
+  getCommentAssetByCommentId: IGetCommentAssetByCommentId;
 }) {
   return async function deleteCommentController(
     httpRequest: Request & { context: { validated: { _id: string } } }
@@ -39,7 +42,13 @@ export default function makeDeleteCommentController({
 
       const comment = await deleteCommentById({ id: comment_id });
 
-      await deleteCommentAssetByCommentId({ comment_id: exists._id });
+      const existing_comment_asset = await getCommentAssetByCommentId({
+        id: exists._id,
+      });
+
+      if (existing_comment_asset) {
+        await deleteCommentAssetByCommentId({ comment_id: exists._id });
+      }
 
       return {
         headers,
